Fix the preference probability calculation
When adding 2 normal distributions, the resulting distribution has a different variance. Code used the variance of the separate distributions as the variance for the added distribution. The preference rating comparison is now updated with the correct distribution. The explanation in the box plot paragraph is also fixed.

Small update in boxplot description